refactor: 合并 MessageCleaner 和 MarkdownNormalizer 为 ContentNormalizer
将入站链中两个独立的 Processor（MessageCleaner priority 30 + MarkdownNormalizer priority 40）
合并为一个 ContentNormalizer（priority 30），消除两个模块之间的隐含契约，简化链配置。

- 新建 content_normalizer.rs，合并飞书 webhook 解析 + markdown 标准化逻辑
- 迁移所有测试到 content_normalizer.rs 和独立测试文件
- 更新 loader.rs 和 mod.rs 中的模块引用
- 删除旧的 message_cleaner.rs、markdown_normalizer.rs 和 message_cleaner_tests.rs

Source: issue #948
Type: refactor
